[gen] refactors in server
