+++ server/package.json
@@ -25,6 +25,7 @@
     "multer": "^2.0.1",
     "nodemon": "^3.1.10",
     "prisma": "^6.10.1",
-    "socket.io": "^4.8.1"
+    "socket.io": "^4.8.1",
+    "stopword": "^3.1.5"
   }
 }
