+++ server/utils/serverWebSocketUtils.js
@@ -1,6 +1,6 @@
 import { Server } from "socket.io";
 import { rollForNotification } from "./notificationsUtils.js";
-import { CONNECTION, DISCONNECT, REQUEST_NOTIFICATION, PING_INTERVAL } from "../constants.js";
+import { CONNECTION, DISCONNECT, REQUEST_NOTIFICATION, PING_INTERVAL } from "./constants.js";
 
 // Creates a websocket server and listens for notification requests
 export const createWebSocket = (server, corsConfig) => {
